micro: extract chat pane send flow
